Abstracting Build Controller methods (partially)
